[Chapter5] made Controller Interface and abtract class for OCP Principle
