+++ src/main/java/webserver/RequestHandler.java
@@ -1,11 +1,9 @@
 package webserver;
 
 import controller.Controller;
-import controller.UserController;
-import db.DataBase;
+import controller.RequestMapping;
 import http.HttpRequest;
 import http.HttpResponse;
-import model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -13,7 +11,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.Socket;
-import java.util.Collection;
 
 public class RequestHandler extends Thread {
     private static final Logger log = LoggerFactory.getLogger(RequestHandler.class);
@@ -33,65 +30,17 @@ public class RequestHandler extends Thread {
             HttpResponse response = new HttpResponse(out);
             String path = getDefaultPath(request.getPath());
 
-            if ("/user/create".equals(path)) {
-                createUser(request, response);
-            } else if ("/user/login".equals(path)) {
-                login(request, response);
-            } else if ("/user/list".equals(path)) {
-                listUser(request, response);
-            } else {
+            Controller controller = RequestMapping.getController(path);
+
+            if(controller == null) {
                 response.forward(path);
+            } else {
+                controller.service(request, response);
             }
+
         } catch (IOException e) {
             log.error(e.getMessage());
         }
-    }
-
-    private void listUser(HttpRequest request, HttpResponse response) {
-        if (!isLogin(request.getCookie("logined"))) {
-            response.sendRedirect("/user/login.html");
-        }
-        Collection<User> users = DataBase.findAll();
-        StringBuilder sb = new StringBuilder();
-        sb.append("<table border='1'>");
-        for (User user : users) {
-            sb.append("<tr>");
-            sb.append("<td>" + user.getUserId() + "</td>");
-            sb.append("<td>" + user.getName() + "</td>");
-            sb.append("<td>" + user.getEmail() + "</td>");
-            sb.append("</tr>");
-        }
-        sb.append("</table>");
-
-        response.forwardBody(sb.toString());
-    }
-
-    private void login(HttpRequest request, HttpResponse response) {
-        User user = DataBase.findUserById(request.getParameter("userId"));
-
-        if (user == null) {
-            response.sendRedirect("/user/login_failed.html");
-            return;
-        }
-
-        if (user.getPassword().equals(request.getParameter("password"))) {
-            response.addHeaders("Set-Cookie", "logined=true;path=/");
-            response.sendRedirect("/index.html");
-        } else {
-            response.sendRedirect("/user/login_failed.html");
-        }
-    }
-
-    private void createUser(HttpRequest request, HttpResponse response) {
-        User user = new User(
-                request.getParameter("userId"),
-                request.getParameter("password"),
-                request.getParameter("name"),
-                request.getParameter("email"));
-        DataBase.addUser(user);
-        log.debug("User :{}", user);
-
-        response.sendRedirect("/index.html");
     }
 
     private String getDefaultPath(String path) {
@@ -99,11 +48,6 @@ public class RequestHandler extends Thread {
             return "/index.html";
         }
         return path;
-    }
-
-    private boolean isLogin(String cookieValue) {
-        if(cookieValue == null) return false;
-        return Boolean.parseBoolean(cookieValue);
     }
 
 }
